+++ CHANGES.md
@@ -2,6 +2,8 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [HEAD]
+
 ## [0.2.0] - 2016-08-15
 ### Added
 - Added `Grove.Buzzer`.
@@ -10,5 +12,5 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Initial release with support for `Grove.Button`.
 
-[Unreleased]: https://github.com/bendiken/nerves_grove/compare/0.2.0...HEAD
+[HEAD]:  https://github.com/bendiken/nerves_grove/compare/0.2.0...HEAD
 [0.2.0]: https://github.com/bendiken/nerves_grove/compare/0.1.0...0.2.0