+++ CHANGES.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [HEAD] - Unreleased
 ### Added
+
+## [0.3.0] - 2016-08-17
+### Added
 - Added `Grove.Buzzer.on/1` and `Grove.Buzzer.off/1`.
 - Added `Grove.OLED.Display`.
 - Added `Grove.Relay`.
@@ -16,5 +19,6 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Initial release with support for `Grove.Button`.
 
-[HEAD]:  https://github.com/bendiken/nerves_grove/compare/0.2.0...HEAD
+[HEAD]:  https://github.com/bendiken/nerves_grove/compare/0.3.0...HEAD
+[0.3.0]: https://github.com/bendiken/nerves_grove/compare/0.2.0...0.3.0
 [0.2.0]: https://github.com/bendiken/nerves_grove/compare/0.1.0...0.2.0